change client side to server side with app/blog/[slug]/edit/page.tsx
